vendor: Support system-as-root A-only backuptool

+++ prebuilt/common/bin/backuptool.sh
@@ -4,8 +4,8 @@
 #
 
 export C=/tmp/backupdir
-export S=/system
-
+export S=$2
+export V=9
 export ADDOND_VERSION=1
 
 # Scripts in /system/addon.d expect to find backuptool.functions in /tmp
@@ -13,8 +13,11 @@ cp -f /tmp/install/bin/backuptool.functions /tmp
 
 # Preserve /system/addon.d in /tmp/addon.d
 preserve_addon_d() {
-  mkdir -p /tmp/addon.d/
-  cp -a /system/addon.d/* /tmp/addon.d/
+  if [ -d $S/addon.d/ ]; then
+    mkdir -p /tmp/addon.d/
+    cp -a $S/addon.d/* /tmp/addon.d/
+    chmod 755 /tmp/addon.d/*.sh
+  fi
 
     # Discard any scripts that aren't at least our version level
     for f in /postinstall/tmp/addon.d/*sh; do
@@ -32,8 +35,11 @@ preserve_addon_d() {
 
 # Restore /system/addon.d in /tmp/addon.d
 restore_addon_d() {
-  cp -a /tmp/addon.d/* /system/addon.d/
-  rm -rf /tmp/addon.d/
+  if [ -d /tmp/addon.d/ ]; then
+    mkdir -p $S/addon.d/
+    cp -a /tmp/addon.d/* $S/addon.d/
+    rm -rf /tmp/addon.d/
+  fi
 }
 
 # Execute /system/addon.d/*.sh scripts with $1 parameter
